+++ actix-tls/CHANGES.md
@@ -1,10 +1,10 @@
 # Changes
 
-[1.0.0-alpha.3] - 2019-12-xx
+[1.0.0-alpha.3] - 2019-12-07
 
 ### Changed
 
-* Migrate to `tokio=0.2.2`
+* Migrate to tokio 0.2
 
 * Enable rustls acceptor service